Adding responsive styling
added a common class 'results' to tip and total for common styling.
Made the most of the element's widths and heights flexible.
Made the body's display flex so the main area can be centered.
Removed the main area's fixed positioning as it is positioned with
    bodys align items
Made the main area's display flex so the direction of the children can
    be easily changed for responsiveness
Added media query styling for mobile